Search for 3D epoxy spheres in median filtered stacks
This macro is the first full working copy.  It contains four search
functions, all designed to look for all spheres in an image.

Algorithm 1) This algorithm is especially good at finding large,
air-filled spheres, as is the only algorithm allowed to find oblate
ellipsoids.  This algorithm searches for spheres assuming that they are
touching the slide.  Specifically, it starts at the top of the stack,
and incrementally goes down looking for any solid object that is equal
to the defined thickness (referred to as a sphere cap).  It then
measures the distance from the tip of the cap to the slide  and assumes
that this is the radius of the sphere.  It then moves to the projected
equator, and measures the median XY radius.  This is important because
some spheres were crushed by the coverslip, and therefore are quite
oblate.  The algorithm then creates the corresponding oblate ellipsoid,
and tests whether the data contained within the predicted volume is an
ellipsoid, the data fills the entire volume, the sphere is not touching
the edge of the image, and that the sphere is not touching a known
sphere. if all this is true, then the sphere is added to the list.

Algorithm 2) This algorithm is especially good at finding spheres that
are not resting on the slide, and is only one of two algorithms that can
find spheres not on the slide.  This algorithm creates an EDT of a MIP
of a segment of the stack, and looks for centroids of a sufficiently
large diameter.  It then verifies candidate spheres by cropping the
stack down to the projected sphere, and creating an EDT of the YZ MIP
(side on projection).  If the XY EDT radius is within a tolerance  range
of the YZ EDT (sphericity check), and the spheres do not touch the edge
of  the image and do not touch a known sphere, then they are added to
the list.

This algorithm is sensitive to artifacts on the slide, so it steps
incrementally down towards the slide to eliminate as many spheres as
possible near the top first.

Algorithm 3) This algorithm is especially good at finding tightly
clustered spheres and very small spheres.  This algorithm is allowed to
find the smallest spheres of all the algorithms.  This algorithm also
assumes all spheres are on the slide (cannot find floating spheres).
This algorithm starts by creating a 3D EDT of a mask without connections
or known spheres.  It then searched up from the largest centroid in the
3D EDT for the top of the sphere in the mask.  Since the EDT mask
expands sphere, the algorithm then switches over to the mask created for
the cap search (algorithm 1) and then searches down form the EDT cap
position, looking for the corresponding cap in the cap search mask.  The
use of the second mask as a verification is critical to proper sphere
volume estimate, and the elimination of false positives.  If the cap has
a sufficiently large radius, then the projected sphere is created, and
if it neither touches the edge of the image or a known sphere, then the
sphere is added to the list.

Algorithm 4) This is the final search algorithm used to find all spheres
above a cutoff volume that were missed by the prior algorithms.  This
algorithm creates a mask with all the known spheres removed, and then
creates a 3D EDT.  It then blurs the EDT (to prevent over segmentation)
and is then segmented via a classical watershed function.  A 3D EDT of
the segmented objects is then also created.  The segmented objects are
the filtered for objects that touch known spheres, do not touch the edge
of the image, and have a sufficiently large radius.  All objects that
meet these three criteria are then checked to make sure they do not
touch known spheres, and are then added to the list.

The macro then creates a results table of the projected spheres and all
the measures parameters, as well as a text based log of all the search
decisions it made while the macro was running, as well as reporting the
total time required for the search.

The macro then color codes the stack containing the found spheres by the
search algorithm that found them (#1 = red, #2 = green, #3 = blue, #4 =
magenta), so that the results can be tuned if a specific algorithm
proves problematic.

The algorithm is also accelerated by removing slices from the stack
whenever possible when performing time consuming steps, and also
performing image math on only selected regions rather than the whole
image.